mbv_africa/requirements.txt
Django
djangorestframework
PyHive
pure-sasl
thrift
thrift-sasl
django-filter
drf-spectacular
djangorestframework-simplejwt
pandas
openpyxl
celery
redis
django-cors-headers
python-decouple
